+++ src/RecordTranslate/RecordInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Efabrica\Translatte\Record;
+
+interface RecordInterface
+{
+    public function save(string $message): void;
+}